+++ packages/assets-controllers/src/TokenRatesController.test.ts
@@ -33,7 +33,11 @@ import type {
   TokenPrice,
   TokenPricesByTokenAddress,
 } from './token-prices-service/abstract-token-prices-service';
-import { controllerName, TokenRatesController } from './TokenRatesController';
+import {
+  controllerName,
+  DEFAULT_CACHE_REFRESH_THRESHOLD,
+  TokenRatesController,
+} from './TokenRatesController';
 import type {
   Token,
   TokenRatesControllerMessenger,
@@ -119,6 +123,10 @@ describe('TokenRatesController', () => {
       await withController(async ({ controller }) => {
         expect(controller.state).toStrictEqual({
           marketData: {},
+          supportedChainIds: {
+            timestamp: 0,
+            data: [],
+          },
         });
       });
     });
@@ -134,6 +142,10 @@ describe('TokenRatesController', () => {
         async ({ controller }) => {
           expect(controller.state).toStrictEqual({
             marketData: {},
+            supportedChainIds: {
+              timestamp: 0,
+              data: [],
+            },
           });
         },
       );
@@ -750,6 +762,10 @@ describe('TokenRatesController', () => {
             options: {
               interval: 100,
               state: {
+                supportedChainIds: {
+                  timestamp: Date.now() + DEFAULT_CACHE_REFRESH_THRESHOLD + 1,
+                  data: [ChainId.mainnet],
+                },
                 marketData: {
                   [ChainId.mainnet]: {
                     '0x02': {
@@ -810,6 +826,10 @@ describe('TokenRatesController', () => {
             options: {
               interval: 100,
               state: {
+                supportedChainIds: {
+                  timestamp: Date.now() + DEFAULT_CACHE_REFRESH_THRESHOLD + 1,
+                  data: [ChainId.mainnet],
+                },
                 marketData: {
                   [ChainId.mainnet]: {
                     '0x02': {
@@ -865,10 +885,16 @@ describe('TokenRatesController', () => {
       });
 
       it('should update exchange rates when network state changes without adding a new network', async () => {
+        const tokenPricesService = buildMockTokenPricesService({
+          fetchSupportedChainIds: jest
+            .fn()
+            .mockResolvedValue([ChainId.mainnet]),
+        });
         await withController(
           {
             options: {
               interval: 100,
+              tokenPricesService,
             },
             mockNetworkClientConfigurationsByNetworkClientId: {
               'AAAA-BBBB-CCCC-DDDD': buildCustomNetworkClientConfiguration({
@@ -1318,7 +1344,11 @@ describe('TokenRatesController', () => {
     describe('start', () => {
       it('should poll and update rate in the right interval', async () => {
         const interval = 100;
-        const tokenPricesService = buildMockTokenPricesService();
+        const tokenPricesService = buildMockTokenPricesService({
+          fetchSupportedChainIds: jest
+            .fn()
+            .mockResolvedValue([ChainId.mainnet]),
+        });
         jest.spyOn(tokenPricesService, 'fetchTokenPrices');
         await withController(
           {
@@ -1365,7 +1395,11 @@ describe('TokenRatesController', () => {
     describe('stop', () => {
       it('should stop polling', async () => {
         const interval = 100;
-        const tokenPricesService = buildMockTokenPricesService();
+        const tokenPricesService = buildMockTokenPricesService({
+          fetchSupportedChainIds: jest
+            .fn()
+            .mockResolvedValue([ChainId.mainnet]),
+        });
         jest.spyOn(tokenPricesService, 'fetchTokenPrices');
         await withController(
           {
@@ -1420,7 +1454,9 @@ describe('TokenRatesController', () => {
 
     it('should poll on the right interval', async () => {
       const interval = 100;
-      const tokenPricesService = buildMockTokenPricesService();
+      const tokenPricesService = buildMockTokenPricesService({
+        fetchSupportedChainIds: jest.fn().mockResolvedValue([ChainId.mainnet]),
+      });
       jest.spyOn(tokenPricesService, 'fetchTokenPrices');
       await withController(
         {
@@ -1465,6 +1501,9 @@ describe('TokenRatesController', () => {
         it('returns the exchange rates directly', async () => {
           const tokenPricesService = buildMockTokenPricesService({
             fetchTokenPrices: fetchTokenPricesWithIncreasingPriceForEachToken,
+            fetchSupportedChainIds: jest
+              .fn()
+              .mockResolvedValue([ChainId.mainnet]),
             validateCurrencySupported(currency: unknown): currency is string {
               return currency === 'ETH';
             },
@@ -1552,6 +1591,10 @@ describe('TokenRatesController', () => {
                     },
                   },
                 },
+                supportedChainIds: {
+                  timestamp: Date.now(),
+                  data: [ChainId.mainnet],
+                },
               });
             },
           );
@@ -1565,6 +1608,9 @@ describe('TokenRatesController', () => {
               .reply(200, { LOL: fallbackRate });
             const tokenPricesService = buildMockTokenPricesService({
               fetchTokenPrices: fetchTokenPricesWithIncreasingPriceForEachToken,
+              fetchSupportedChainIds: jest
+                .fn()
+                .mockResolvedValue([ChainId.mainnet]),
               validateCurrencySupported(currency: unknown): currency is string {
                 return currency !== 'LOL';
               },
@@ -1671,7 +1717,11 @@ describe('TokenRatesController', () => {
                 new Error('market does not exist for this coin pair'),
               );
 
-            const tokenPricesService = buildMockTokenPricesService();
+            const tokenPricesService = buildMockTokenPricesService({
+              fetchSupportedChainIds: jest
+                .fn()
+                .mockResolvedValue([ChainId.mainnet]),
+            });
             await withController(
               {
                 options: {
@@ -1725,7 +1775,11 @@ describe('TokenRatesController', () => {
 
       it('should stop polling', async () => {
         const interval = 100;
-        const tokenPricesService = buildMockTokenPricesService();
+        const tokenPricesService = buildMockTokenPricesService({
+          fetchSupportedChainIds: jest
+            .fn()
+            .mockResolvedValue([ChainId.mainnet]),
+        });
         jest.spyOn(tokenPricesService, 'fetchTokenPrices');
         await withController(
           {
@@ -1818,7 +1872,13 @@ describe('TokenRatesController', () => {
       });
 
       it('does not update state if there are no tokens for the given chain', async () => {
+        const tokenPricesService = buildMockTokenPricesService({
+          fetchSupportedChainIds: jest
+            .fn()
+            .mockResolvedValue([toHex(2), ChainId.mainnet]),
+        });
         await withController(
+          { options: { tokenPricesService } },
           async ({
             controller,
             triggerTokensStateChange,
@@ -1850,6 +1910,10 @@ describe('TokenRatesController', () => {
             });
 
             expect(controller.state).toStrictEqual({
+              supportedChainIds: {
+                timestamp: Date.now(),
+                data: [toHex(2), ChainId.mainnet],
+              },
               marketData: {
                 [ChainId.mainnet]: {
                   '0x0000000000000000000000000000000000000000': {
@@ -1864,6 +1928,9 @@ describe('TokenRatesController', () => {
 
       it('does not update state if the price update fails', async () => {
         const tokenPricesService = buildMockTokenPricesService({
+          fetchSupportedChainIds: jest
+            .fn()
+            .mockResolvedValue([ChainId.mainnet]),
           fetchTokenPrices: jest
             .fn()
             .mockRejectedValue(new Error('Failed to fetch')),
@@ -1913,6 +1980,9 @@ describe('TokenRatesController', () => {
           .sort();
         const tokenPricesService = buildMockTokenPricesService({
           fetchTokenPrices: fetchTokenPricesWithIncreasingPriceForEachToken,
+          fetchSupportedChainIds: jest
+            .fn()
+            .mockResolvedValue([ChainId.mainnet, 0x3e7]),
         });
         const fetchTokenPricesSpy = jest.spyOn(
           tokenPricesService,
@@ -1976,6 +2046,9 @@ describe('TokenRatesController', () => {
           '0x0000000000000000000000000000000000000003',
         ];
         const tokenPricesService = buildMockTokenPricesService({
+          fetchSupportedChainIds: jest
+            .fn()
+            .mockResolvedValue([ChainId.mainnet]),
           fetchTokenPrices: jest.fn().mockResolvedValue({
             [tokenAddresses[0]]: {
               currency: 'ETH',
@@ -2038,29 +2111,31 @@ describe('TokenRatesController', () => {
               selectedNetworkClientId: InfuraNetworkType.mainnet,
             });
 
-            expect(controller.state).toMatchInlineSnapshot(`
-                        Object {
-                          "marketData": Object {
-                            "0x1": Object {
-                              "0x0000000000000000000000000000000000000001": Object {
-                                "currency": "ETH",
-                                "tokenAddress": "0x0000000000000000000000000000000000000001",
-                                "value": 0.001,
-                              },
-                              "0x0000000000000000000000000000000000000002": Object {
-                                "currency": "ETH",
-                                "tokenAddress": "0x0000000000000000000000000000000000000002",
-                                "value": 0.002,
-                              },
-                              "0x0000000000000000000000000000000000000003": Object {
-                                "currency": "ETH",
-                                "tokenAddress": "0x0000000000000000000000000000000000000003",
-                                "value": 0.003,
-                              },
-                            },
-                          },
-                        }
-                    `);
+            expect(controller.state).toStrictEqual({
+              supportedChainIds: {
+                timestamp: Date.now(),
+                data: ['0x1'],
+              },
+              marketData: {
+                '0x1': {
+                  '0x0000000000000000000000000000000000000001': {
+                    currency: 'ETH',
+                    tokenAddress: '0x0000000000000000000000000000000000000001',
+                    value: 0.001,
+                  },
+                  '0x0000000000000000000000000000000000000002': {
+                    currency: 'ETH',
+                    tokenAddress: '0x0000000000000000000000000000000000000002',
+                    value: 0.002,
+                  },
+                  '0x0000000000000000000000000000000000000003': {
+                    currency: 'ETH',
+                    tokenAddress: '0x0000000000000000000000000000000000000003',
+                    value: 0.003,
+                  },
+                },
+              },
+            });
           },
         );
       });
@@ -2072,6 +2147,7 @@ describe('TokenRatesController', () => {
             '0x0000000000000000000000000000000000000002',
           ];
           const tokenPricesService = buildMockTokenPricesService({
+            fetchSupportedChainIds: jest.fn().mockResolvedValue([toHex(2)]),
             fetchTokenPrices: jest.fn().mockResolvedValue({
               [tokenAddresses[0]]: {
                 currency: 'ETH',
@@ -2120,24 +2196,28 @@ describe('TokenRatesController', () => {
                 setChainAsCurrent: false,
               });
 
-              expect(controller.state).toMatchInlineSnapshot(`
-                            Object {
-                              "marketData": Object {
-                                "0x2": Object {
-                                  "0x0000000000000000000000000000000000000001": Object {
-                                    "currency": "ETH",
-                                    "tokenAddress": "0x0000000000000000000000000000000000000001",
-                                    "value": 0.001,
-                                  },
-                                  "0x0000000000000000000000000000000000000002": Object {
-                                    "currency": "ETH",
-                                    "tokenAddress": "0x0000000000000000000000000000000000000002",
-                                    "value": 0.002,
-                                  },
-                                },
-                              },
-                            }
-                        `);
+              expect(controller.state).toStrictEqual({
+                supportedChainIds: {
+                  timestamp: Date.now(),
+                  data: ['0x2'],
+                },
+                marketData: {
+                  '0x2': {
+                    '0x0000000000000000000000000000000000000001': {
+                      currency: 'ETH',
+                      tokenAddress:
+                        '0x0000000000000000000000000000000000000001',
+                      value: 0.001,
+                    },
+                    '0x0000000000000000000000000000000000000002': {
+                      currency: 'ETH',
+                      tokenAddress:
+                        '0x0000000000000000000000000000000000000002',
+                      value: 0.002,
+                    },
+                  },
+                },
+              });
             },
           );
         });
@@ -2155,6 +2235,7 @@ describe('TokenRatesController', () => {
           '0x0000000000000000000000000000000000000002',
         ];
         const tokenPricesService = buildMockTokenPricesService({
+          fetchSupportedChainIds: jest.fn().mockResolvedValue([toHex(137)]),
           fetchTokenPrices: jest.fn().mockResolvedValue({
             [tokenAddresses[0]]: {
               currency: 'ETH',
@@ -2220,38 +2301,40 @@ describe('TokenRatesController', () => {
             });
 
             // token value in terms of matic should be (token value in eth) * (eth value in matic)
-            expect(controller.state).toMatchInlineSnapshot(`
-                        Object {
-                          "marketData": Object {
-                            "0x89": Object {
-                              "0x0000000000000000000000000000000000000001": Object {
-                                "allTimeHigh": undefined,
-                                "allTimeLow": undefined,
-                                "currency": "UNSUPPORTED",
-                                "dilutedMarketCap": undefined,
-                                "high1d": undefined,
-                                "low1d": undefined,
-                                "marketCap": undefined,
-                                "price": 0.0005,
-                                "tokenAddress": "0x0000000000000000000000000000000000000001",
-                                "totalVolume": undefined,
-                              },
-                              "0x0000000000000000000000000000000000000002": Object {
-                                "allTimeHigh": undefined,
-                                "allTimeLow": undefined,
-                                "currency": "UNSUPPORTED",
-                                "dilutedMarketCap": undefined,
-                                "high1d": undefined,
-                                "low1d": undefined,
-                                "marketCap": undefined,
-                                "price": 0.001,
-                                "tokenAddress": "0x0000000000000000000000000000000000000002",
-                                "totalVolume": undefined,
-                              },
-                            },
-                          },
-                        }
-                    `);
+            expect(controller.state).toStrictEqual({
+              supportedChainIds: {
+                timestamp: Date.now(),
+                data: ['0x89'],
+              },
+              marketData: {
+                '0x89': {
+                  '0x0000000000000000000000000000000000000001': {
+                    allTimeHigh: undefined,
+                    allTimeLow: undefined,
+                    currency: 'UNSUPPORTED',
+                    dilutedMarketCap: undefined,
+                    high1d: undefined,
+                    low1d: undefined,
+                    marketCap: undefined,
+                    price: 0.0005,
+                    tokenAddress: '0x0000000000000000000000000000000000000001',
+                    totalVolume: undefined,
+                  },
+                  '0x0000000000000000000000000000000000000002': {
+                    allTimeHigh: undefined,
+                    allTimeLow: undefined,
+                    currency: 'UNSUPPORTED',
+                    dilutedMarketCap: undefined,
+                    high1d: undefined,
+                    low1d: undefined,
+                    marketCap: undefined,
+                    price: 0.001,
+                    tokenAddress: '0x0000000000000000000000000000000000000002',
+                    totalVolume: undefined,
+                  },
+                },
+              },
+            });
           },
         );
       });
@@ -2267,6 +2350,7 @@ describe('TokenRatesController', () => {
           .map(buildAddress)
           .sort();
         const tokenPricesService = buildMockTokenPricesService({
+          fetchSupportedChainIds: jest.fn().mockResolvedValue([toHex(999)]),
           fetchTokenPrices: fetchTokenPricesWithIncreasingPriceForEachToken,
           validateCurrencySupported: (
             currency: unknown,
@@ -2361,6 +2445,9 @@ describe('TokenRatesController', () => {
           '0x0000000000000000000000000000000000000002',
         ];
         const tokenPricesService = buildMockTokenPricesService({
+          fetchSupportedChainIds: jest
+            .fn()
+            .mockResolvedValue([ChainId.mainnet]),
           fetchTokenPrices: jest.fn().mockResolvedValue({
             [tokenAddresses[0]]: {
               currency: 'ETH',
@@ -2373,9 +2460,6 @@ describe('TokenRatesController', () => {
               value: 0.002,
             },
           }),
-          validateChainIdSupported(_chainId: unknown): _chainId is Hex {
-            return false;
-          },
         });
         await withController(
           {
@@ -2417,22 +2501,25 @@ describe('TokenRatesController', () => {
               selectedNetworkClientId,
             });
 
-            expect(controller.state).toMatchInlineSnapshot(`
-                        Object {
-                          "marketData": Object {
-                            "0x3e7": Object {
-                              "0x0000000000000000000000000000000000000001": undefined,
-                              "0x0000000000000000000000000000000000000002": undefined,
-                            },
-                          },
-                        }
-                      `);
+            expect(controller.state).toStrictEqual({
+              supportedChainIds: {
+                timestamp: Date.now(),
+                data: ['0x1'],
+              },
+              marketData: {
+                '0x3e7': {
+                  '0x0000000000000000000000000000000000000001': undefined,
+                  '0x0000000000000000000000000000000000000002': undefined,
+                },
+              },
+            });
           },
         );
       });
 
       it('correctly calls the Price API with unqiue native token addresses (e.g. MATIC)', async () => {
         const tokenPricesService = buildMockTokenPricesService({
+          fetchSupportedChainIds: jest.fn().mockResolvedValue(['0x89']),
           fetchTokenPrices: jest.fn().mockResolvedValue({
             '0x0000000000000000000000000000000000001010': {
               currency: 'MATIC',
@@ -2499,6 +2586,7 @@ describe('TokenRatesController', () => {
         });
         const tokenPricesService = buildMockTokenPricesService({
           fetchTokenPrices: fetchTokenPricesMock,
+          fetchSupportedChainIds: jest.fn().mockResolvedValue([toHex(1)]),
         });
         await withController(
           { options: { tokenPricesService } },
@@ -2540,24 +2628,26 @@ describe('TokenRatesController', () => {
 
             expect(fetchTokenPricesMock).toHaveBeenCalledTimes(1);
 
-            expect(controller.state).toMatchInlineSnapshot(`
-                        Object {
-                          "marketData": Object {
-                            "0x1": Object {
-                              "0x0000000000000000000000000000000000000001": Object {
-                                "currency": "ETH",
-                                "tokenAddress": "0x0000000000000000000000000000000000000001",
-                                "value": 0.001,
-                              },
-                              "0x0000000000000000000000000000000000000002": Object {
-                                "currency": "ETH",
-                                "tokenAddress": "0x0000000000000000000000000000000000000002",
-                                "value": 0.002,
-                              },
-                            },
-                          },
-                        }
-                    `);
+            expect(controller.state).toStrictEqual({
+              supportedChainIds: {
+                timestamp: Date.now(),
+                data: ['0x1'],
+              },
+              marketData: {
+                '0x1': {
+                  '0x0000000000000000000000000000000000000001': {
+                    currency: 'ETH',
+                    tokenAddress: '0x0000000000000000000000000000000000000001',
+                    value: 0.001,
+                  },
+                  '0x0000000000000000000000000000000000000002': {
+                    currency: 'ETH',
+                    tokenAddress: '0x0000000000000000000000000000000000000002',
+                    value: 0.002,
+                  },
+                },
+              },
+            });
           },
         );
       });
@@ -2583,7 +2673,17 @@ describe('TokenRatesController', () => {
           fetchTokenPrices: fetchTokenPricesMock,
         });
         await withController(
-          { options: { tokenPricesService } },
+          {
+            options: {
+              tokenPricesService,
+              state: {
+                supportedChainIds: {
+                  timestamp: Date.now() + DEFAULT_CACHE_REFRESH_THRESHOLD,
+                  data: ['0x1'],
+                },
+              },
+            },
+          },
           async ({
             controller,
             triggerTokensStateChange,
@@ -2650,6 +2750,64 @@ describe('TokenRatesController', () => {
           },
         );
       });
+
+      it('will not call fetchSupportedChainIds if the cache is valid', async () => {
+        const tokenAddresses = [
+          '0x0000000000000000000000000000000000000001',
+          '0x0000000000000000000000000000000000000002',
+        ];
+        const fetchTokenPricesMock = jest.fn().mockResolvedValue({
+          [tokenAddresses[0]]: {
+            currency: 'ETH',
+            tokenAddress: tokenAddresses[0],
+            value: 0.001,
+          },
+          [tokenAddresses[1]]: {
+            currency: 'ETH',
+            tokenAddress: tokenAddresses[1],
+            value: 0.002,
+          },
+        });
+        const fetchSupportedChainIdsMock = jest.fn().mockResolvedValue(['0x1']);
+        const tokenPricesService = buildMockTokenPricesService({
+          fetchTokenPrices: fetchTokenPricesMock,
+          fetchSupportedChainIds: fetchSupportedChainIdsMock,
+        });
+        await withController(
+          {
+            options: {
+              tokenPricesService,
+            },
+            mockTokensControllerState: {
+              allTokens: {
+                '0x1': {
+                  [defaultSelectedAddress]: [
+                    {
+                      address: mockTokenAddress,
+                      decimals: 0,
+                      symbol: '',
+                      aggregators: [],
+                    },
+                  ],
+                },
+              },
+            },
+          },
+          async ({ controller }) => {
+            // use promise.all to call the updateExchangeRates method twice
+            await Promise.all([
+              controller.updateExchangeRates([
+                { chainId: ChainId.mainnet, nativeCurrency: 'ETH' },
+              ]),
+              controller.updateExchangeRates([
+                { chainId: ChainId.mainnet, nativeCurrency: 'ETH' },
+              ]),
+            ]);
+
+            expect(fetchSupportedChainIdsMock).toHaveBeenCalledTimes(1);
+          },
+        );
+      });
     });
   });
 
@@ -2682,6 +2840,10 @@ describe('TokenRatesController', () => {
             },
           },
         },
+        supportedChainIds: {
+          timestamp: 0,
+          data: [],
+        },
       };
 
       await withController(
@@ -2697,6 +2859,10 @@ describe('TokenRatesController', () => {
 
           expect(controller.state).toStrictEqual({
             marketData: {},
+            supportedChainIds: {
+              timestamp: 0,
+              data: [],
+            },
           });
         },
       );
@@ -2739,6 +2905,10 @@ describe('TokenRatesController', () => {
         ).toMatchInlineSnapshot(`
           Object {
             "marketData": Object {},
+            "supportedChainIds": Object {
+              "data": Array [],
+              "timestamp": 0,
+            },
           }
         `);
       });
@@ -2755,6 +2925,10 @@ describe('TokenRatesController', () => {
         ).toMatchInlineSnapshot(`
           Object {
             "marketData": Object {},
+            "supportedChainIds": Object {
+              "data": Array [],
+              "timestamp": 0,
+            },
           }
         `);
       });
@@ -2999,8 +3173,8 @@ function buildMockTokenPricesService(
     async fetchExchangeRates() {
       return {};
     },
-    validateChainIdSupported(_chainId: unknown): _chainId is Hex {
-      return true;
+    async fetchSupportedChainIds() {
+      return [];
     },
     validateCurrencySupported(_currency: unknown): _currency is string {
       return true;